Removed the geometry pass system
This is in favour of giving the renderer more control over how the scene it rendered
