--- clove/include/Clove/Rendering/RenderPasses/SkinningPass.hpp
@@ -1,30 +1,0 @@
-#pragma once
-
-#include "Clove/Rendering/RenderPasses/GeometryPass.hpp"
-
-#include <memory>
-
-namespace clove {
-    class GhaComputePipelineObject;
-    class GhaFactory;
-}
-
-namespace clove {
-    class SkinningPass : public GeometryPass {
-        //FUNCTIONS
-    public:
-        SkinningPass();
-
-        SkinningPass(SkinningPass const &other) = delete;
-        SkinningPass(SkinningPass &&other) noexcept;
-
-        SkinningPass &operator=(SkinningPass const &other) = delete;
-        SkinningPass &operator=(SkinningPass &&other) noexcept;
-
-        ~SkinningPass();
-        
-        Id getId() const override;
-
-        void execute(RenderGraph &renderGraph, PassData const &passData) override;
-    };
-}
