--- clove/include/Clove/Rendering/Techniques/ForwardLightingTechnique.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "Clove/Rendering/Techniques/Technique.hpp"
-
-namespace clove {
-    Technique createForwardLightingTechnique();
-    Technique createSkinnedForwardLightingTechnique();
-}
