+++ clove/source/SubSystems/RenderSubSystem.cpp
@@ -55,15 +55,9 @@ namespace clove {
         entityManager->forEach([this](TransformComponent const &transform, StaticModelComponent const &staticModel) {
             if(staticModel.model.isValid()) {
                 mat4f const modelTransform{ transform.worldMatrix };
-                std::array<mat4f, MAX_JOINTS> matrixPalet{};
-                matrixPalet.fill(mat4f{ 1.0f });
-
-                std::set<GeometryPass::Id> passIds;
-                for(auto const &technique : staticModel.model->getTechniques()) {
-                    passIds.insert(technique.passIds.begin(), technique.passIds.end());
-                }
+                
                 for(auto const &mesh : staticModel.model->getMeshes()) {
-                    renderer->submitMesh(Renderer::MeshInfo{ mesh, staticModel.material, modelTransform, matrixPalet, passIds });
+                    renderer->submitMesh(Renderer::MeshInfo{ mesh, staticModel.material, modelTransform });
                 }
             }
         });
@@ -73,12 +67,8 @@ namespace clove {
                 mat4f const modelTransform{ transform.worldMatrix };
                 auto const matrixPalet{ animatedModel.animator.update(deltaTime) };
 
-                std::set<GeometryPass::Id> passIds;
-                for(auto const &technique : animatedModel.model->getTechniques()) {
-                    passIds.insert(technique.passIds.begin(), technique.passIds.end());
-                }
                 for(auto const &mesh : animatedModel.model->getMeshes()) {
-                    renderer->submitMesh(Renderer::MeshInfo{ mesh, animatedModel.material, modelTransform, matrixPalet, passIds });
+                    renderer->submitMesh(Renderer::MeshInfo{ mesh, animatedModel.material, modelTransform, matrixPalet });
                 }
             }
         });
